Description of trigramanalyzer_wordlanguageid.py
This program is meant to find the character trigrams, based on their positions (single, pre, mid and post) in two languages.
It calculates the probability of occurrence of the character trigrams in the two languages by reading text in the languages from specified text files which must be saved in UTF-8. The probability of occurrence of trigrams in each language equals to co-efficient of discrimination/identification.
In addition, it obtains the trigrams that are common to the two languages. The co-efficient of discrimination/identification for the trigrams that are common to the two languages becomes zero.